python_service/requirements.txt
fastapi
uvicorn
pytesseract
pillow
langchain
ollama
langchain_community
pdf2image
python-multipart
langchain-ollama
opencv-python
numpy
